# Refactor tests for private sites and Lockout component