--- a/security/class-lockout.php
+++ b/security/class-lockout.php
@@ -43,9 +43,9 @@ class Lockout {
 		if ( defined( 'VIP_LOCKOUT_STATE' ) ) {
 			$user = wp_get_current_user();
 
-			switch ( VIP_LOCKOUT_STATE ) {
+			switch ( constant( 'VIP_LOCKOUT_STATE' ) ) {
 				case 'warning':
-					$show_notice = apply_filters( 'vip_lockout_show_notice', $user->has_cap( 'manage_options' ), VIP_LOCKOUT_STATE, $user );
+					$show_notice = apply_filters( 'vip_lockout_show_notice', $user->has_cap( 'manage_options' ), constant( 'VIP_LOCKOUT_STATE' ), $user );
 					if ( $show_notice ) {
 						$this->render_warning_notice();
 
@@ -55,7 +55,7 @@ class Lockout {
 					break;
 
 				case 'locked':
-					$show_notice = apply_filters( 'vip_lockout_show_notice', $user->has_cap( 'edit_posts' ), VIP_LOCKOUT_STATE, $user );
+					$show_notice = apply_filters( 'vip_lockout_show_notice', $user->has_cap( 'edit_posts' ), constant( 'VIP_LOCKOUT_STATE' ), $user );
 					if ( $show_notice ) {
 						$this->render_locked_notice();
 
@@ -76,7 +76,7 @@ class Lockout {
 		$seen_warning = get_user_meta( $user->ID, self::USER_SEEN_WARNING_KEY, true );
 
 		if ( ! $seen_warning ) {
-			add_user_meta( $user->ID, self::USER_SEEN_WARNING_KEY, VIP_LOCKOUT_STATE, true );
+			add_user_meta( $user->ID, self::USER_SEEN_WARNING_KEY, constant( 'VIP_LOCKOUT_STATE' ), true );
 			// phpcs:ignore WordPress.DateTime.RestrictedFunctions.date_date -- not sure if it is safe to replace with gmdate()
 			add_user_meta( $user->ID, self::USER_SEEN_WARNING_TIME_KEY, date( 'Y-m-d H:i:s' ), true );
 		}
@@ -87,7 +87,7 @@ class Lockout {
 		<div id="lockout-warning" class="notice-warning wrap clearfix" style="align-items: center;background: #ffffff;border-left-width:4px;border-left-style:solid;border-radius: 6px;display: flex;margin-top: 30px;padding: 30px;line-height: 2em;">
 			<div class="dashicons dashicons-warning" style="display:flex;float:left;margin-right:2rem;font-size:38px;align-items:center;margin-left:-20px;color:#ffb900;"></div>
 			<div style="display: flex;align-items: center;" >
-				<h3><?php echo wp_kses_post( VIP_LOCKOUT_MESSAGE ); ?></h3>
+				<h3><?php echo wp_kses_post( constant( 'VIP_LOCKOUT_MESSAGE' ) ); ?></h3>
 			</div>
 		</div>
 		<?php
@@ -98,7 +98,7 @@ class Lockout {
 		<div id="lockout-warning" class="notice-error wrap clearfix" style="align-items: center;background: #ffffff;border-left-width:4px;border-left-style:solid;border-radius: 6px;display: flex;margin-top: 30px;padding: 30px;line-height: 2em;">
 			<div class="dashicons dashicons-warning" style="display:flex;float:left;margin-right:2rem;font-size:38px;align-items:center;margin-left:-20px;color:#dc3232;"></div>
 			<div style="display: flex;align-items: center;" >
-				<h3><?php echo wp_kses_post( VIP_LOCKOUT_MESSAGE ); ?></h3>
+				<h3><?php echo wp_kses_post( constant( 'VIP_LOCKOUT_MESSAGE' ) ); ?></h3>
 			</div>
 		</div>
 		<?php
@@ -117,7 +117,7 @@ class Lockout {
 	 * @return array
 	 */
 	public function filter_user_has_cap( $user_caps, $caps, $args, $user ) {
-		if ( defined( 'VIP_LOCKOUT_STATE' ) && 'locked' === VIP_LOCKOUT_STATE ) {
+		if ( defined( 'VIP_LOCKOUT_STATE' ) && 'locked' === constant( 'VIP_LOCKOUT_STATE' ) ) {
 			if ( is_automattician( $user->ID ) ) {
 				return $user_caps;
 			}
@@ -146,7 +146,7 @@ class Lockout {
 	 * @return  array
 	 */
 	public function filter_site_admin_option( $pre_option ) {
-		if ( defined( 'VIP_LOCKOUT_STATE' ) && 'locked' === VIP_LOCKOUT_STATE ) {
+		if ( defined( 'VIP_LOCKOUT_STATE' ) && 'locked' === constant( 'VIP_LOCKOUT_STATE' ) ) {
 			if ( is_automattician() ) {
 				return $pre_option;
 			}
@@ -167,7 +167,7 @@ class Lockout {
 	 * Instead, just block updates to the option if a site is locked.
 	 */
 	public function filter_prevent_site_admin_option_updates( $value, $old_value ) {
-		if ( defined( 'VIP_LOCKOUT_STATE' ) && 'locked' === VIP_LOCKOUT_STATE ) {
+		if ( defined( 'VIP_LOCKOUT_STATE' ) && 'locked' === constant( 'VIP_LOCKOUT_STATE' ) ) {
 			return $old_value;
 		}