--- a/tests/bootstrap.php
+++ b/tests/bootstrap.php
@@ -25,6 +25,7 @@ function _manually_load_plugin() {
 
 	require_once __DIR__ . '/../async-publish-actions.php';
 	require_once __DIR__ . '/../performance.php';
+
 	require_once __DIR__ . '/../security.php';
 
 	require_once __DIR__ . '/../schema.php';
@@ -113,6 +114,7 @@ switch ( getenv( 'WPVIP_PARSELY_INTEGRATION_TEST_MODE' ) ) {
 		break;
 }
 
+require_once __DIR__ . '/mock-constants.php';
 require_once __DIR__ . '/mock-header.php';
 
 require $_tests_dir . '/includes/bootstrap.php';