This PR refactors tests for Private Sites and Lockouts and Warnings by mocking all calls to `defined()` and `constant()`; this allows us to run all those tests in a single process.

For this to work, I had to patch `security/class-private-sites.php` and `security/class-lockout.php` to use `constant()` instead of accessing constants directly.

Mocking relies upon the [name resolution rules](http://php.net/manual/en/language.namespaces.rules.php) of PHP namespaces. However, because the PHP compiler tries to optimize away all calls to `defined()`, it is crucial that the mock is loaded before the tested class.

Timings (PS + Lockout)
* Before: 23.92 seconds + 31.89 seconds
* After: 2.68 seconds + 3.53 seconds